SystemCode/backend/requirements.txt
fastapi
uvicorn
ultralytics
pillow
torch
numpy
python-multipart